set up controller for creating new order
