+++ server/DB/controllers/newOrder.js
@@ -1,0 +1,28 @@
+import uuidv4 from 'uuid/v4';
+import db from '../db/index';
+
+const newOrder = (req, res) => {
+  const text = `INSERT INTO
+      orders(id, foodname, quantity, price, foodstatus, owner_id, date_created, date_modified)
+      VALUES($1, $2, $3, $4, $5, $6, $7, $8)
+      returning *`;
+  const values = [
+    uuidv4(),
+    req.body.foodname,
+    req.body.quantity,
+    req.body.price,
+    req.body.foodstatus,
+    req.body.owner_id,
+    new Date(),
+    new Date(),
+  ];
+
+  try {
+    const { rows } = db.query(text, values);
+    return res.status(201).send(rows[0]);
+  } catch (error) {
+    return res.status(400).send(error);
+  }
+};
+
+export default newOrder;
