+++ server/DB/controllers/orders.js
@@ -1,51 +1,13 @@
-import moment from 'moment';
-import uuidv4 from 'uuid/v4';
 import db from '../db';
 
 const Order = {
 
-  async create(req, res) {
-    const text = `INSERT INTO
-      orders(id, foodname, quantity, price, foodstatus, owner_id, date_created, date_modified)
-      VALUES($1, $2, $3, $4, $5, $6, $7, $8)
-      returning *`;
-    const values = [
-      uuidv4(),
-      req.body.foodname,
-      req.body.quantity,
-      req.body.price,
-      req.body.foodstatus,
-      req.body.owner_id,
-      moment(new Date()),
-      moment(new Date()),
-    ];
-
-    try {
-      const { rows } = await db.query(text, values);
-      return res.status(201).send(rows[0]);
-    } catch (error) {
-      return res.status(400).send(error);
-    }
-  },
 
   async getAll(req, res) {
     const findAllQuery = 'SELECT * FROM orders';
     try {
       const { rows, rowCount } = await db.query(findAllQuery);
       return res.status(200).send({ rows, rowCount });
-    } catch (error) {
-      return res.status(400).send(error);
-    }
-  },
-
-  async getOne(req, res) {
-    const text = 'SELECT * FROM orders WHERE id = $1';
-    try {
-      const { rows } = await db.query(text, [req.params.id]);
-      if (!rows[0]) {
-        return res.status(404).send({ message: 'reflection not found' });
-      }
-      return res.status(200).send(rows[0]);
     } catch (error) {
       return res.status(400).send(error);
     }
@@ -65,7 +27,7 @@ const Order = {
         req.body.success || rows[0].success,
         req.body.low_point || rows[0].low_point,
         req.body.take_away || rows[0].take_away,
-        moment(new Date()),
+        new Date(),
         req.params.id,
       ];
       const response = await db.query(updateOneQuery, values);
